Move more to CRCommandFlow
