+++ Sources/CrunchyrollSwift-DL/CRDownload.swift
@@ -19,26 +19,12 @@ struct CRDownload: ParsableCommand {
                 if parsed.type == .series, let url = URL(string: url) {
                     print("Getting seiresId from web page")
                     if let seriesId = CRWebParser.seriesId(url) {
-                        if let collections = CRAPIHelper.getCollections(sessionId, seriesId) {
-                            print("\nChoose a colletion:")
-                            for (index, collection) in collections.enumerated() {
-                                print("\(index + 1): \(collection.name)")
-                            }
-                            let collectionChoise = UserInteract.chooseNumber(from: 1...collections.count)
-                            let selectedCollection = collections[collectionChoise-1]
-                            print("Getting episodes from \(collectionChoise): \(selectedCollection.name)")
+                        if let selectedCollection = CRCommandFlow.selectCollection(sessionId, seriesId) {
                             guard let selectedCollectionId = Int(selectedCollection.id) else {
                                 print("Collection id is invaildate")
                                 continue
                             }
-                            if let episodes = CRAPIHelper.getMedias(sessionId, selectedCollectionId) {
-                                print("\nChoose a episode:")
-                                for (index, episode) in episodes.enumerated() {
-                                    print("\(index + 1): \(episode.name ?? "Untitled")")
-                                }
-                                let episodeChoise = UserInteract.chooseNumber(from: 1...episodes.count)
-                                let selectedEpisode = episodes[episodeChoise-1]
-                                print("Getting info from \(episodeChoise): \(selectedEpisode.name ?? "Untitled")")
+                            if let selectedEpisode = CRCommandFlow.selectEpisode(sessionId, selectedCollectionId) {
                                 guard let selectedMediaId = Int(selectedEpisode.id) else {
                                     print("Episode id is invaildate")
                                     continue
